Frame the page
Created a "header" which has the Player One and Player Two names. Player
One will always have tile 'O' and Player Two will always have tile 'X'.

Left a space in the middle for the actual game board. I think I'll make
that div be a grid and have the first row be 1fr and the second be 3fr.
The first row will be a status bar showing W/L/T for each player and the
current round that the game is on.

There's also a footer on the bottom that links back to the GitHub repo
for this project.