Add & Style round stats bar and Play! button
Added a button between the player names that'll start the game and
disable the player names from being edited.

Added the round stats bar which will be populated with numbers once the
game starts. It's currently hidden but below the round stats bar there
is a div with 3 child divs inside it for the actual board itself which
will then be populated once the Play! button is clicked.