See description of this commit
I got the visuals of the page improved a little more and added the
boxes on the left and right that'll eventually denote who's turn it is
but as of right now they're statically colored red.

I also implemented the buttons for the most part. The turns switch
properly but I still need to figure out how I want to get the colors to
change based on who's turn it is.

The big issue right now is the `detectWin()` function. As of right now
it detects horizontal wins, doesn't detect vertical, and diagonal is
untested. I'm not sure why it's not detecting vertical wins though
which is the biggest issue.

Will look more in depth at the `detectWin()` function later.